requirements.txt
streamlit
python-dotenv
langchain
langchain-openai
langchain-community
google-api-python-client
google-auth
google-auth-oauthlib
openai
os